+++ apps/front/src/lib/util.ts
@@ -1,0 +1,3 @@
+export function isNotNull<T>(w: T | null): w is T {
+  return w !== null;
+}