+++ apps/front/src/search/SearchTab.tsx
@@ -1,11 +1,4 @@
-import {
-  Button,
-  GeolocationOption,
-  LazyAutocomplete,
-  Select,
-  SimpleTooltip,
-  useCopyToClipboard,
-} from "pentatrion-design";
+import { Button, GeolocationOption, LazyAutocomplete, Select } from "pentatrion-design";
 import { useAppDispatch, useAppSelector } from "../store";
 import { searchFeatureChanged, selectSearchFeature } from "./searchSlice";
 import {
@@ -14,32 +7,25 @@ import {
   GeoPointOption,
   c2cWaypointSearch,
   createGeolocationFeature,
-  getCoordsStr,
   ignSearch,
-  stringifyGeoOption,
 } from "pentatrion-geo";
 import {
   SearchEngine,
-  coordsUnitChanged,
   searchEngineChanged,
   searchEngines,
-  selectCoordsUnit,
   selectSearchEngine,
   selectViewState,
-  tabChanged,
 } from "../store/mapSlice";
 import { useNotification } from "pentatrion-design/redux";
 import { useT } from "talkr";
-import { useCallback, useMemo, useState } from "react";
-import { directionWayPointsAddedFromSearch } from "~/direction/directionSlice";
-import Isochrone from "./Isochrone";
+import { useMemo } from "react";
 import { inputSearchDebounceDelay } from "~/config/constants";
 import { SearchEngineOption, StarOption } from "~/components/search-engine/SearchEngineOption";
 import { SearchEngineSelection } from "~/components/search-engine/SearchEngineSelection";
 import { iconBySearchEngine } from "~/components/search-engine/util";
 import { activationChanged } from "~/geolocation/geolocationSlice";
-
-type Action = "isochrone" | "direction" | "raw";
+import FeatureInfos from "./FeatureInfos";
+import GeolocationInfos from "~/geolocation/GeolocationInfos";
 
 export default function SearchTab() {
   const searchFeature = useAppSelector(selectSearchFeature);
@@ -47,22 +33,7 @@ export default function SearchTab() {
   const viewState = useAppSelector(selectViewState);
   const { notifyError } = useNotification();
   const searchEngine = useAppSelector(selectSearchEngine);
-  const coordsUnit = useAppSelector(selectCoordsUnit);
   const { T } = useT();
-  const [, copy] = useCopyToClipboard();
-  const { notify } = useNotification();
-
-  const [action, setAction] = useState<Action | null>(null);
-  const setOrToggleAction = useCallback(
-    (a: Action) => {
-      if (a === action) {
-        setAction(null);
-      } else {
-        setAction(a);
-      }
-    },
-    [action],
-  );
 
   const searchEngineOptions = useMemo<StarOption[]>(() => {
     return searchEngines.map((s) => ({
@@ -107,7 +78,7 @@ export default function SearchTab() {
                 dispatch(activationChanged(true));
               }}
             >
-              <i className="fe-locate"></i>
+              <i className="fe-geolocation"></i>
             </Button>
           }
           selection={searchFeature}
@@ -130,115 +101,8 @@ export default function SearchTab() {
           }}
         />
       </div>
-      {searchFeature && (
-        <>
-          {searchFeature.type === "Feature" && (
-            <div>
-              <div className="setting">
-                <div className="text-hint">{T("coordinates")}</div>
-                <div>
-                  <Button
-                    className="size-small text-hint"
-                    variant="ghost"
-                    color="weak"
-                    onClick={() => dispatch(coordsUnitChanged())}
-                  >
-                    {T(`coordsUnit.${coordsUnit}`)}{" "}
-                  </Button>
-                  &nbsp;
-                  <span
-                    className="can-copy"
-                    onClick={() => {
-                      const value = getCoordsStr(searchFeature.geometry.coordinates, coordsUnit);
-                      copy(value);
-                      notify(`${T("copiedIntoClipboard")} : ${value}`);
-                    }}
-                  >
-                    {getCoordsStr(searchFeature.geometry.coordinates, coordsUnit)}
-                  </span>
-                </div>
-              </div>
-              <div className="setting">
-                <div className="text-hint">{T("elevation")}</div>
-                <div>
-                  {searchFeature.properties.originalProperties?.elevation ??
-                    searchFeature.geometry.coordinates[2] ??
-                    "-"}
-                  <span className="text-hint"> m</span>
-                </div>
-              </div>
-              {["city", "postcode"].map((key) => {
-                const value = searchFeature.properties.originalProperties?.[key] as
-                  | string
-                  | undefined;
-                if (!value) {
-                  return null;
-                }
-                return (
-                  <div className="setting" key={key}>
-                    <div className="text-hint">{T(`property.${key}`)}</div>
-                    <div>{value}</div>
-                  </div>
-                );
-              })}
-            </div>
-          )}
-          <div className="actions">
-            {searchFeature.type !== "geolocation" && (
-              <SimpleTooltip content={T("tooltip.direction")} placement="top-start">
-                <Button
-                  variant="light"
-                  color="weak"
-                  onClick={() => {
-                    dispatch(directionWayPointsAddedFromSearch(searchFeature));
-                    dispatch(tabChanged("direction"));
-                  }}
-                >
-                  <i className="fe-route"></i>
-                </Button>
-              </SimpleTooltip>
-            )}
-            <SimpleTooltip content={T("tooltip.code")} placement="top">
-              <Button
-                variant="light"
-                color="weak"
-                selected={action === "raw"}
-                onClick={() => setOrToggleAction("raw")}
-              >
-                <i className="fe-code"></i>
-              </Button>
-            </SimpleTooltip>
-            <SimpleTooltip content={T("tooltip.isochrone")} placement="top-end">
-              <Button
-                variant="light"
-                color="weak"
-                selected={action === "isochrone"}
-                onClick={() => setOrToggleAction("isochrone")}
-              >
-                <i className="fe-isochrone"></i>
-              </Button>
-            </SimpleTooltip>
-          </div>
-
-          {action === "raw" && (
-            <>
-              <div className="separator"></div>
-              <textarea
-                className="ll-textarea text-sm raw"
-                readOnly
-                defaultValue={stringifyGeoOption(searchFeature)}
-              />
-            </>
-          )}
-
-          {action === "isochrone" && (
-            <>
-              <div className="separator"></div>
-              <Isochrone />
-            </>
-          )}
-        </>
-      )}
+      {searchFeature?.type === "Feature" && <FeatureInfos />}
+      {searchFeature?.type === "geolocation" && <GeolocationInfos />}
     </div>
   );
 }