+++ packages/pentatrion-design/components/quick-settings/QuickSettings.stories.tsx
@@ -65,6 +65,26 @@ export const Basic = () => {
         </Button>
       </div>
 
+      <div className="ll-steps-extra">
+        <Button variant="ghost" color="weak">
+          <span
+            className="ll-marker"
+            style={{ "--marker-color": "#c0c0c0", "--marker-size": "34px" }}
+          >
+            <span className="marker">
+              <span className="ovale"></span>
+              <i className="fe-plus"></i>
+            </span>
+            <span className="target"></span>
+          </span>
+          <span>Ajouter un point</span>
+        </Button>
+
+        <Button icon variant="text" color="weak">
+          <i className="fe-sliders"></i>
+        </Button>
+      </div>
+
       <div className="setting">
         <div>Temps</div>
         <Input suffix="min" defaultValue="0" />
@@ -102,12 +122,46 @@ export const Basic = () => {
           }}
         ></Select>
       </div>
-      <div className="setting constraints">
+      <div className="setting">
         <div>Passages autorisés</div>
         <div className="ll-input-checkbox-container placement-inline">
           <Checkbox checked={true}>Péages</Checkbox>
           <Checkbox checked={false}>Ponts</Checkbox>
         </div>
+      </div>
+
+      <div className="setting multiple">
+        <div>Éviter</div>
+        <div className="ll-input-checkbox-container placement-block">
+          <Checkbox checked={true}>
+            <span>Sections à péage</span>
+          </Checkbox>
+          <Checkbox checked={false}>
+            <span>Ponts</span>
+          </Checkbox>
+          <Checkbox checked={false}>
+            <span>Tunnels</span>
+          </Checkbox>
+        </div>
+      </div>
+
+      <div className="separator"></div>
+
+      <div>
+        <div className="setting">
+          <div>Distance</div>
+          <div>
+            325 <span className="text-hint">km</span>
+          </div>
+        </div>
+        <div className="setting">
+          <div>Durée</div>
+          <div>3h 25min</div>
+        </div>
+      </div>
+
+      <div className="align-right">
+        <Button>Calculer</Button>
       </div>
     </div>
   );