+++ packages/pentatrion-geo/src/components/autocomplete/AutocompleteGeoOption.tsx
@@ -1,7 +1,7 @@
 import { memo, useId } from "react";
 import { useListItem } from "@floating-ui/react";
 import cn from "classnames";
-import { GeoOption, GeoPointOption, GeometryWithCoordinates, getTypeLabel } from "../..";
+import { GeoPointOption, getTypeLabel } from "../..";
 import { GeolocationOption, useAutocomplete } from "pentatrion-design";
 import { getValue } from "pentatrion-design/components/autocomplete/util";